+++ src/components/OverflowMenu/index.js
@@ -1,0 +1,1 @@
+export { OverflowMenu } from 'carbon-components-react';
