+++ src/components/OverflowMenuItem/index.js
@@ -1,0 +1,1 @@
+export { OverflowMenuItem } from 'carbon-components-react';
